Revert "Gen 3: Reduce bundle size (#3405)"
This reverts commit 2b50529a73b95398e08bbbca55c9b65f63cf2ef9.
