+++ src/util/CountryUtil.js
@@ -10,8 +10,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import _ from 'underscore';
-import { loc } from './loc';
+import { _, loc } from '@okta/courage';
 import bundles from './Bundles';
 import countryCallingCodes from './countryCallingCodes';
 const fn = {};
