+++ src/util/Util.js
@@ -12,8 +12,7 @@
  */
 
 /* eslint complexity: [2, 13], max-depth: [2, 3] */
-import _ from 'underscore';
-import { loc } from './loc';
+import { _, loc } from '@okta/courage';
 import Enums from './Enums';
 import Logger from './Logger';
 import BrowserFeatures from './BrowserFeatures';
