+++ src/v2/ion/IonResponseHelper.js
@@ -10,8 +10,8 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import _ from 'underscore';
-import { getMessage, getMessageKey } from './i18nUtils';
+import { _ } from '@okta/courage';
+import { getMessage, getMessageKey } from './i18nTransformer';
 
 const convertErrorMessageToErrorSummary = (formName, remediationValues = []) => {
   return _.chain(remediationValues)
