+++ src/v3/jest.config.js
@@ -49,7 +49,9 @@ module.exports = {
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/src/tests/__mocks__/fileMock.js',
     '^.+\\.svg$': '<rootDir>/svgMockTransformer.js',
     '\\.(css|less|scss)$': 'identity-obj-proxy',
+    '^@okta/courage$': `${PACKAGES}/@okta/courage-dist/esm/src/index.js`,
     '^@okta/okta-i18n-bundles$': `${PROJECT_ROOT}/src/util/Bundles`,
+    '^@okta/qtip$': `${PACKAGES}/@okta/qtip2/dist/jquery.qtip.js`,
     '^@okta/mocks/(.*)': `${PROJECT_ROOT}/playground/mocks/$1`,
     'util/Logger': `${PROJECT_ROOT}/src/util/Logger`,
     'util/Bundles': `${PROJECT_ROOT}/src/util/Bundles`,
