+++ src/v3/src/components/WebAuthNSubmitButton/WebAuthNSubmitButton.test.tsx
@@ -22,7 +22,7 @@ import {
 
 import WebAuthNSubmitButton from './WebAuthNSubmitButton';
 
-jest.mock('../../../../v2/ion/i18nUtils', () => ({
+jest.mock('../../../../v2/ion/i18nTransformer', () => ({
   getMessageFromBrowserError: (error: Error) => (error.name === 'NotAllowedError' ? 'Operation not allowed' : error.message),
 }));
