+++ src/v3/src/transformer/i18n/util.ts
@@ -15,7 +15,7 @@ import { IdxTransaction } from '@okta/okta-auth-js';
 import {
   getI18nKey,
   getI18NParams,
-} from '../../../../v2/ion/i18nUtils';
+} from '../../../../v2/ion/i18nTransformer';
 import {
   FieldElement, TokenReplacement, UISchemaElement,
 } from '../../types';
