+++ src/v3/src/transformer/layout/consent/transformEnduserConsent.ts
@@ -10,7 +10,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import { doesI18NKeyExist } from '../../../../../v2/ion/i18nUtils';
+import { doesI18NKeyExist } from '../../../../../v2/ion/i18nTransformer';
 import {
   ButtonElement,
   ButtonType,
