+++ src/v3/src/transformer/password/passwordSettingsUtils.test.ts
@@ -14,7 +14,7 @@ import { PASSWORD_REQUIREMENTS_KEYS } from '../../constants';
 import { ComplexityRequirements } from '../../types';
 import { getComplexityItems } from './passwordSettingsUtils';
 
-jest.mock('util/loc', () => ({
+jest.mock('@okta/courage', () => ({
   loc: jest.fn().mockImplementation(
     (key) => key,
   ),
