+++ src/v3/src/transformer/terminal/transformUnhandledErrors.ts
@@ -12,7 +12,7 @@
 
 import { AuthApiError, OAuthError } from '@okta/okta-auth-js';
 
-import { getMessage } from '../../../../v2/ion/i18nUtils';
+import { getMessage } from '../../../../v2/ion/i18nTransformer';
 import {
   FormBag,
   InfoboxElement,
