+++ src/v3/tsconfig.base.json
@@ -16,11 +16,17 @@
     "moduleResolution": "node",
     "noEmit": true,
     "paths": {
+      "@okta/courage": [
+        "../../packages/@okta/courage-dist"
+      ],
       "@okta/mocks/*": [
         "../../playground/mocks/*"
       ],
       "@okta/okta-i18n-bundles": [
         "../util/Bundles.ts"
+      ],
+      "@okta/qtip": [
+        "../../packages/@okta/qtip2/dist/jquery.qtip.js"
       ],
       "config/config.json": [
         "../config/config.json"
